Strip punctuation from beginning of vcc function names.
In some cases vcc function names is parsed with a punctuation in front.
This breaks autocomplete in vscode. my_goto_backend.backend() completes
as my_goto_backend..backend() etc.

This change removes the punctuation from the start of function names if
there is one before adding it to the list of vcc functions.